// src/pages/app/provider/common-files/enums.ts
export enum DisableDate {
    FUTURE = 'future',
    PAST = 'past'
}

export enum AccountsType {
    Patient = "Patient",
    Insurance = "Insurance",
}

export enum Options {
    days = '30days',
    month = 'month',
    quarterly = 'quarterly'
}

export enum TableType {
    ROS = 'Review of System',
    PE = 'Physical Exam'
}